requirements.txt
pandas==2.2.2
pyodbc==5.0.1
SQLAlchemy==2.0.32
python-dotenv==1.0.1